+++ src/site/elements/pages/rester-page-request.js
@@ -8,6 +8,7 @@ import '../../../../node_modules/@polymer/iron-form/iron-form.js';
 import '../../../../node_modules/@polymer/iron-icon/iron-icon.js';
 import '../../../../node_modules/@polymer/iron-media-query/iron-media-query.js';
 import '../../../../node_modules/@polymer/iron-pages/iron-pages.js';
+import '../../../../node_modules/@polymer/paper-button/paper-button.js';
 import '../../../../node_modules/@polymer/paper-icon-button/paper-icon-button.js';
 import '../../../../node_modules/@polymer/paper-item/paper-item.js';
 import '../../../../node_modules/@polymer/paper-listbox/paper-listbox.js';
@@ -964,7 +965,11 @@ class RESTerPageRequest extends RESTerLintMixin(
             .catch((error) => {
                 this._setRequestIsSending(false);
                 if (error.name !== 'AbortError') {
-                    this.showError(error);
+                    this.showError(error, null, [
+                        'Is the URL correct?',
+                        'Is your network connection working?',
+                        'Could this be caused by an invalid HTTPS certificate? If so, try to open the URL in a new tab and accept the certificate warning. Then try the request in RESTer again.',
+                    ]);
                 }
             });
     }